B4: signedness-audit fixes (98.071%) (#17542)
* gbaque: fix signedness in LoadAll spModeFlags compare (96.74%)

The m_spModeFlags[i] != 0 guard compiled to an unsigned cmplwi; the
target uses a signed cmpwi on the int-promoted byte. Cast the read to
int so mwcc emits lbz + cmpwi (no extsb), matching the original.


* cflat_r2class: fix signedness in onClassSystemFunc case -0x48 (96.45%)

The case -0x48 guard object->m_localBase[0] != 0 compiled to an
unsigned cmplwi (m_localBase is unsigned int*); the target uses signed
cmpwi. Cast to int to match, consistent with sibling cases -0x8C/-0x4F.


---------

+++ src/cflat_r2class.cpp
@@ -1556,7 +1556,7 @@ int CFlatRuntime2::onClassSystemFunc(CFlatRuntime::CObject* object, int, int com
 		}
 		case -0x48:
 			engineObject->m_stepSlopeLimit = reinterpret_cast<float*>(object->m_localBase)[1];
-			if (object->m_localBase[0] != 0) {
+			if (static_cast<int>(object->m_localBase[0]) != 0) {
 				engineObject->m_lookAtTimer = engineObject->m_stepSlopeLimit;
 			}
 			PushValue(this, object, 0);
